% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/detrend.R
\name{detrend}
\alias{detrend}
\title{Detrend a time series by subtracting Theil-Sen slope}
\usage{
detrend(x)
}
\arguments{
\item{x}{numeric, a time series
return time series detrended}
}
\description{
Detrend a time series by subtracting Theil-Sen slope
}
\examples{
y=c(2,3,5,6)
detrend(y)

}
\author{
Ken Butler, \email{butler@utsc.utoronto.ca}
}
