% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kendall_S.R
\name{kendall_S}
\alias{kendall_S}
\title{Calculate S statistic for Kendall correlation with time (+1 for concordance, -1 for discordance)}
\usage{
kendall_S(x)
}
\arguments{
\item{x}{numeric, a time series}
}
\value{
number of concordances minus number of discordances
}
\description{
Calculate S statistic for Kendall correlation with time (+1 for concordance, -1 for discordance)
}
\examples{

kendall_S(c(2,3,5,4))

}
\author{
Ken Butler, \email{butler@utsc.utoronto.ca}
}
