% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kendall_Z.R
\name{kendall_Z}
\alias{kendall_Z}
\title{Kendall test statistic for testing time trend (no autocorrelation)}
\usage{
kendall_Z(x)
}
\arguments{
\item{x}{numeric, a time series}
}
\value{
test statistic
}
\description{
Kendall test statistic for testing time trend (no autocorrelation)
}
\examples{
kendall_Z(c(2,3,5,4))

}
\author{
Ken Butler, \email{butler@utsc.utoronto.ca}
}
