% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kendall_Z_adjusted.R
\name{kendall_Z_adjusted}
\alias{kendall_Z_adjusted}
\title{Mann-Kendall correlation test, unadjusted and adjusted for autocorrelation, with P-values}
\usage{
kendall_Z_adjusted(x, ...)
}
\arguments{
\item{x}{numeric, a time series}

\item{...}{optional arguments to "acf", eg. lag.max (see sample_size_ratio)}
}
\value{
list: unadjusted and adjusted Z statistics, effective sample size ratio, unadjusted and adjusted P-values
}
\description{
Mann-Kendall correlation test, unadjusted and adjusted for autocorrelation, with P-values
}
\examples{


set.seed(457299)
x=rnorm(100) # uncorrelated (no adjustment needed)
x=x+0.02*(1:100) # with trend
kendall_Z_adjusted(x)

# ar(1) data
set.seed(457298)
xx=arima.sim(list(ar=0.8),100) # autocorrelated, needs adjusting
kendall_Z_adjusted(xx)
# P-value adjusted is much less significant

}
\author{
Ken Butler, \email{butler@utsc.utoronto.ca}
}
