% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sample_size_ratio.R
\name{sample_size_ratio}
\alias{sample_size_ratio}
\title{ratio of actual to effective sample sizes for possibly autocorrelated time series}
\usage{
sample_size_ratio(x, ...)
}
\arguments{
\item{x}{numeric, a time series}

\item{..., }{optional parameters for acf()}
}
\value{
sample size ratio (a number)
}
\description{
ratio of actual to effective sample sizes for possibly autocorrelated time series
}
\examples{
set.seed(457298)
xx=stats::arima.sim(list(ar=0.8),100) # posoitively autocorrelated, sample size ratio > 1
sample_size_ratio(xx)

}
\author{
Ken Butler, \email{butler@utsc.utoronto.ca}
}
