% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/var_kendall_S.R
\name{var_kendall_S}
\alias{var_kendall_S}
\title{Variance of Kendall correlation S statistic under null hypothesis of no trend and assuming independence}
\usage{
var_kendall_S(x)
}
\arguments{
\item{x}{numeric, a time series (only its length is used)}
}
\description{
Variance of Kendall correlation S statistic under null hypothesis of no trend and assuming independence
}
\examples{
var_kendall_S(rnorm(100))

}
\author{
Ken Butler, \email{butler@utsc.utoronto.ca}
}
