% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sum_df.R
\name{sum_df}
\alias{sum_df}
\title{display calculations for sum in adjustment for autocorrelation (mainly for debugging)}
\usage{
sum_df(x, ...)
}
\arguments{
\item{x}{numeric, a time series}

\item{..., }{optional arguments for acf}
}
\value{
data frame showing terms in sum
}
\description{
display calculations for sum in adjustment for autocorrelation (mainly for debugging)
}
\author{
Ken Butler, \email{butler@utsc.utoronto.ca}
}
